Refactor(Entity): 기본 엔티티 수정
- ERD 변경으로 인해 내부 속성(컬럼)을 변경
- LocalDateTime -> LocalTime 변경
- 필요한 속성 추가 (ex. deleted,등)
- User -> 여러 설정 값을 받아서 사용자 편의 향상 예정
